Add implicit operator for list of exceptions
The commit includes the addition of an implicit operator in both Result.cs and Error.cs classes that accepts a list of exceptions. The GetResultType method in Result.cs was also refactored to a property named ResultType. Unit tests were added to confirm the correct handling of multiple exceptions. Also, OkTests.cs was renamed to GenericOkTests.cs and its class name was correspondingly updated.